Organize file structure to separate ORM refactor
